🎨 Improved members table structure and removed low-signal tests
ref https://linear.app/ghost/issue/BER-3470/member-list-fails-to-render-in-chrome-for-massive-member-counts
The members table work had accumulated too much local complexity, so this folds the table chrome into a dedicated helper, removes dead plumbing, and drops an arithmetic helper test that was not adding meaningful confidence beyond browser coverage.
